+++ trunk/code/kmeans_omp.cpp
@@ -1,14 +1,14 @@
 #include "kmeans.h"
 #include "kmeans_omp.h"
 
-int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize) {
+int kmeans_omp(uchar *particle_data, float *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize) {
     reset_timer( 1 );
     reset_timer( 2 );
     reset_timer( 3 );
     reset_timer( 4 );
 
     // assign centers
-    double *cluster_sizes = new double[cluster_count];
+    int *cluster_sizes = new int[cluster_count];
     start_timer( 4 );
     if (select_centerspp_omp(particle_data, particle_count, dimensions, cluster_count, centers) != 0 ) {
         printf("Error selecting centers");
@@ -29,10 +29,10 @@ int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int di
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) { // TODO parallel for
             
             // find closest center
-            double min_dist = std::numeric_limits<double>::max();
+            float min_dist = std::numeric_limits<float>::max();
             uchar cluster_assignment = std::numeric_limits<uchar>::max();
             for (int center_iter = 0; center_iter < cluster_count; center_iter++) {
-                double dist = compute_distance(particle_data, particle_iter, centers, center_iter, dimensions);
+                float dist = compute_distance(particle_data, particle_iter, centers, center_iter, dimensions);
                 if (dist < min_dist) {
                     min_dist = dist;
                     cluster_assignment = (uchar)center_iter;
@@ -54,7 +54,7 @@ int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int di
             // initialize centers to 0
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<double>(centers, cluster_iter, dim_iter, dimensions) = 0;
+                    array_store<float>(centers, cluster_iter, dim_iter, dimensions) = 0;
                 }
             }
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
@@ -66,7 +66,7 @@ int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int di
                 int cluster_assignment = (int)assignments[particle_iter];
                 cluster_sizes[cluster_assignment]++;
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<double>(centers, cluster_assignment, dim_iter, dimensions) += (double)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
+                    array_store<float>(centers, cluster_assignment, dim_iter, dimensions) += (float)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
 
                 }
             }
@@ -75,8 +75,8 @@ int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int di
             #pragma omp parallel for
             for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {
                 for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                    array_store<double>(centers, cluster_iter, dim_iter, dimensions) =
-                        floor(array_load<double>(centers, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5);
+                    array_store<float>(centers, cluster_iter, dim_iter, dimensions) =
+                        floor(array_load<float>(centers, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5);
                 }
             }
             #pragma omp barrier
@@ -99,8 +99,8 @@ int kmeans_omp(uchar *particle_data, double *centers, int particle_count, int di
 /**
  * Smarter cluster select method
  */
-int select_centerspp_omp(uchar *particle_data, int particle_count, int dimensions, int cluster_count, double *centers) {
-    double *particle_distr = new double[particle_count];
+int select_centerspp_omp(uchar *particle_data, int particle_count, int dimensions, int cluster_count, float *centers) {
+    float *particle_distr = new float[particle_count];
     int distr_length = RAND_MAX;
 
     reset_timer(4);
@@ -111,7 +111,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
     // Select the first center randomly
     int particle_select = rand() % particle_count;
     for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-        array_store<double>(centers, 0, dim_iter, dimensions) = (double)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
+        array_store<float>(centers, 0, dim_iter, dimensions) = (float)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
     }
 
     // Select the other n - 1 centers
@@ -121,9 +121,9 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
         #pragma omp parallel for
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) { // TODO parallel for
             // find closest center based on previously calculated centers
-            double min_dist = std::numeric_limits<double>::max();
+            float min_dist = std::numeric_limits<float>::max();
             for (int center_dist_iter = 0; center_dist_iter < center_iter; center_dist_iter++) {
-                double dist = compute_distance(particle_data, particle_iter, centers, center_dist_iter, dimensions);
+                float dist = compute_distance(particle_data, particle_iter, centers, center_dist_iter, dimensions);
                 if (dist < min_dist) {
                     min_dist = dist;
                 }
@@ -151,7 +151,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
 
         start_timer(6);
         // Select new center
-        double distr_ratio = particle_distr[particle_count-1] / distr_length;
+        float distr_ratio = particle_distr[particle_count-1] / distr_length;
         int select_pos = rand() % distr_length;
         #pragma omp parallel for
         for (int particle_iter = 0; particle_iter < particle_count; particle_iter++) {  // TODO search - make binary, parallelize (note break, tricky)
@@ -164,7 +164,7 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
 
         // Copy new center to array
         for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-            array_store<double>(centers, center_iter, dim_iter, dimensions) = (double)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
+            array_store<float>(centers, center_iter, dim_iter, dimensions) = (float)array_load<uchar>(particle_data, particle_select, dim_iter, dimensions);
         }
     }
     delete [] particle_distr;
@@ -183,10 +183,10 @@ int select_centerspp_omp(uchar *particle_data, int particle_count, int dimension
  * Maybe should be square root of above?
  */
 /*
-double compute_distance(uchar *particle_data, const int particle_iter, double *centers, const int center_iter, const int dimensions) {
-    double dist = 0;
+float compute_distance(uchar *particle_data, const int particle_iter, float *centers, const int center_iter, const int dimensions) {
+    float dist = 0;
     for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-        dist += pow( (double)array_store<uchar>(particle_data, particle_iter, dim_iter, dimensions) - array_load<double>(centers, center_iter, dim_iter, dimensions), 2 );
+        dist += pow( (float)array_store<uchar>(particle_data, particle_iter, dim_iter, dimensions) - array_load<float>(centers, center_iter, dim_iter, dimensions), 2 );
     }
     return dist;
 }
