+++ trunk/code/kmeans_tbb.h
@@ -19,7 +19,7 @@ using namespace tbb;
 class KmeansStep1 {
     uchar *particle_data;
     uchar *assignments;
-    double *centers;
+    float *centers;
     const int cluster_count;
     const int dimensions;
     bool *assignment_change;
@@ -27,10 +27,10 @@ public:
     void operator()( const blocked_range<size_t>& r ) const {
         for( size_t particle_iter=r.begin(); particle_iter!=r.end(); ++particle_iter ) {
             // find closest center
-            double min_dist = std::numeric_limits<double>::max();
+            float min_dist = std::numeric_limits<float>::max();
             uchar cluster_assignment = std::numeric_limits<uchar>::max();
             for (int center_iter = 0; center_iter < cluster_count; center_iter++) {
-                double dist = compute_distance(particle_data, particle_iter, centers, center_iter, dimensions);
+                float dist = compute_distance(particle_data, particle_iter, centers, center_iter, dimensions);
                 if (dist < min_dist) {
                     min_dist = dist;
                     cluster_assignment = (uchar)center_iter;
@@ -44,37 +44,37 @@ public:
             }
         }
     }
-    KmeansStep1(uchar *_particle_data, double *_centers, int _cluster_count, uchar *_assignments, int _dimensions, bool *_assignment_change) : 
+    KmeansStep1(uchar *_particle_data, float *_centers, int _cluster_count, uchar *_assignments, int _dimensions, bool *_assignment_change) : 
                 particle_data(_particle_data), cluster_count(_cluster_count), centers(_centers), assignments(_assignments), dimensions(_dimensions), assignment_change(_assignment_change) {}
 };
 
 class KmeansStep2a {
     uchar *particle_data;
-    double *centers;
+    float *centers;
     uchar *assignments;
     const int cluster_count;
     const int dimensions;
 public:
-    double *sumCenters;
-    double *sumClusterSizes;
+    float *sumCenters;
+    int *sumClusterSizes;
     void operator() (const blocked_range<size_t>& r) const {
         for( size_t particle_iter=r.begin(); particle_iter!=r.end(); ++particle_iter ) {
             int cluster_assignment = (int)assignments[particle_iter];
             sumClusterSizes[cluster_assignment]++;
             for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                array_store<double>(sumCenters, cluster_assignment, dim_iter, dimensions) += (double)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
+                array_store<float>(sumCenters, cluster_assignment, dim_iter, dimensions) += (float)array_load<uchar>(particle_data, particle_iter, dim_iter, dimensions);
             }
         }
     }
 
     KmeansStep2a( KmeansStep2a &x, split ) : 
         particle_data(x.particle_data), centers(x.centers), assignments(x.assignments), cluster_count(x.cluster_count), dimensions(x.dimensions) {
-        sumCenters = new double[cluster_count*dimensions];
-        sumClusterSizes = new double[cluster_count];
+        sumCenters = new float[cluster_count*dimensions];
+        sumClusterSizes = new int[cluster_count];
         for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {    
             sumClusterSizes[cluster_iter] = 0;
             for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                array_store<double>(sumCenters, cluster_iter, dim_iter, dimensions) = 0;
+                array_store<float>(sumCenters, cluster_iter, dim_iter, dimensions) = 0;
             }
         }
     }
@@ -83,19 +83,19 @@ public:
         for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {    
             sumClusterSizes[cluster_iter] += y.sumClusterSizes[cluster_iter];
             for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                array_store<double>(sumCenters, cluster_iter, dim_iter, dimensions) += (double)array_load<double>(y.sumCenters, cluster_iter, dim_iter, dimensions);
+                array_store<float>(sumCenters, cluster_iter, dim_iter, dimensions) += (float)array_load<float>(y.sumCenters, cluster_iter, dim_iter, dimensions);
             }
         }
     }
 
-    KmeansStep2a(uchar *_particle_data, double *_centers, uchar *_assignments, int _cluster_count, int _dimensions) :
+    KmeansStep2a(uchar *_particle_data, float *_centers, uchar *_assignments, int _cluster_count, int _dimensions) :
         particle_data(_particle_data), centers(_centers), assignments(_assignments), cluster_count(_cluster_count), dimensions(_dimensions) {
-        sumCenters = new double[_cluster_count*_dimensions];
-        sumClusterSizes = new double[_cluster_count];
+        sumCenters = new float[_cluster_count*_dimensions];
+        sumClusterSizes = new int[_cluster_count];
         for (int cluster_iter = 0; cluster_iter < cluster_count; cluster_iter++) {    
             sumClusterSizes[cluster_iter] = 0;
             for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                array_store<double>(sumCenters, cluster_iter, dim_iter, dimensions) = 0;
+                array_store<float>(sumCenters, cluster_iter, dim_iter, dimensions) = 0;
             }
         }
     }
@@ -107,25 +107,25 @@ public:
 };            
 
 class KmeansStep2b {
-    double *centers;
-    double *center_sums;
-    double *cluster_sizes;
+    float *centers;
+    float *center_sums;
+    int *cluster_sizes;
     int dimensions;
 public:
     void operator() (const blocked_range<size_t>& r) const {
         for (size_t cluster_iter = r.begin(); cluster_iter != r.end(); ++cluster_iter) {
             for (int dim_iter = 0; dim_iter < dimensions; dim_iter++) {
-                array_store<double>(centers, cluster_iter, dim_iter, dimensions) =
-                    floor(array_load<double>(center_sums, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5);
+                array_store<float>(centers, cluster_iter, dim_iter, dimensions) =
+                    floor(array_load<float>(center_sums, cluster_iter, dim_iter, dimensions)/cluster_sizes[cluster_iter] + 0.5);
             }
         }
         return;
     }
-    KmeansStep2b(double *_centers, double *_center_sums, double *_cluster_sizes, int _dimensions) :
+    KmeansStep2b(float *_centers, float *_center_sums, int *_cluster_sizes, int _dimensions) :
         centers(_centers), center_sums(_center_sums), cluster_sizes(_cluster_sizes), dimensions(_dimensions) {}
 };            
 
-int kmeans_tbb( uchar *data, double *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize, int thread_count );
+int kmeans_tbb( uchar *data, float *centers, int particle_count, int dimensions, int cluster_count, uchar *assignments, int grainSize, int thread_count );
 
 #endif
